utils file refactored
