+++ track_syscalls/syscallfilter/syscallfilter.go
@@ -20,15 +20,46 @@ func Load() (Allowlist, error) {
 	return allowlist, err
 }
 
-func Write(syscalls map[string][]int) error {
-	data := Allowlist{
-		Dependencies: syscalls,
+func Write(newSyscalls map[string][]int) error {
+	var existingAllowlist Allowlist
+
+	// Read existing allowlist if it exists
+	data, err := os.ReadFile("allowlist.json")
+	if err == nil {
+		json.Unmarshal(data, &existingAllowlist)
 	}
-	jsonData, err := json.MarshalIndent(data, "", "  ")
+
+	// Merge new syscalls with existing ones
+	if existingAllowlist.Dependencies == nil {
+		existingAllowlist.Dependencies = make(map[string][]int)
+	}
+	for pkg, syscalls := range newSyscalls {
+		existingAllowlist.Dependencies[pkg] = mergeSyscalls(existingAllowlist.Dependencies[pkg], syscalls)
+	}
+
+	// Write merged data back to file
+	jsonData, err := json.MarshalIndent(existingAllowlist, "", "  ")
 	if err != nil {
 		return err
 	}
 	return os.WriteFile("allowlist.json", jsonData, 0644)
+}
+
+func mergeSyscalls(existing, new []int) []int {
+	merged := make(map[int]bool)
+	for _, syscall := range existing {
+		merged[syscall] = true
+	}
+	for _, syscall := range new {
+		merged[syscall] = true
+	}
+
+	result := make([]int, 0, len(merged))
+	for syscall := range merged {
+		result = append(result, syscall)
+	}
+	sort.Ints(result)
+	return result
 }
 
 func (a *Allowlist) SyscallAllowed(callerPkg string, syscall int) bool {
@@ -45,12 +76,13 @@ func (a *Allowlist) SyscallAllowed(callerPkg string, syscall int) bool {
 }
 
 func ConvertSyscallsMap(syscalls map[string]map[int]bool) map[string][]int {
-	result := make(map[string][]int)
-	for pkg, calls := range syscalls {
-		for call := range calls {
-			result[pkg] = append(result[pkg], call)
+	convertedSyscalls := make(map[string][]int)
+	for pkg, syscallMap := range syscalls {
+		syscallList := make([]int, 0, len(syscallMap))
+		for syscall := range syscallMap {
+			syscallList = append(syscallList, syscall)
 		}
-		sort.Ints(result[pkg])
+		convertedSyscalls[pkg] = syscallList
 	}
-	return result
+	return convertedSyscalls
 }
